web/requirements_dev.txt
ipython==5.1.0
pytest-django==3.0.0
responses==0.5.1
pytest-cov==2.3.1
